Changed CLI Structure (Non-Breaking), MySQL Exporter uses now the unique_id_generator, Manifest includes License file!
