+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include Pipfile
+include Pipfile LICENSE
